fix: movind prepend stuff into Extension
