+++ src/ZhorteinSymfonyToolboxBundle.php
@@ -14,36 +14,6 @@ use Zhortein\SymfonyToolboxBundle\DependencyInjection\ZhorteinSymfonyToolboxExte
 
 class ZhorteinSymfonyToolboxBundle extends AbstractBundle
 {
-    public function prependExtension(ContainerConfigurator $container, ContainerBuilder $builder): void
-    {
-        if (!$this->isAssetMapperAvailable($builder)) {
-            return;
-        }
-
-        $builder->prependExtensionConfig('framework', [
-            'asset_mapper' => [
-                'paths' => [
-                    __DIR__.'/../assets/dist' => '@zhortein/symfony-toolbox-bundle',
-                ],
-            ],
-        ]);
-    }
-
-    private function isAssetMapperAvailable(ContainerBuilder $container): bool
-    {
-        if (!interface_exists(AssetMapperInterface::class)) {
-            return false;
-        }
-
-        // check that FrameworkBundle 6.3 or higher is installed
-        $bundlesMetadata = $container->getParameter('kernel.bundles_metadata');
-        if (!isset($bundlesMetadata['FrameworkBundle'])) {
-            return false;
-        }
-
-        return is_file($bundlesMetadata['FrameworkBundle']['path'].'/Resources/config/asset_mapper.php');
-    }
-
     /**
      * @param array<int|string, mixed> $config
      *
